handle error edit not exist user

+++ src/main/java/services/ProductService.java
@@ -70,15 +70,21 @@ public class ProductService extends AbstractService implements ModelService {
         if (newModel instanceof ProductModel) {
             ProductModel model = (ProductModel) newModel;
             try (Stream<String> stream = Files.lines(Paths.get(FILE_NAME))) {
+                AtomicBoolean isNotIdExist = new AtomicBoolean(true);
                 List<String> products = new ArrayList<>();
                 stream.map(this::convertModel)
                         .forEach(productModel -> {
                             if (productModel.getId().equals(model.getId())) {
+                                isNotIdExist.set(false);
                                 products.add(getAsRawData(model));
                             } else {
                                 products.add(getAsRawData(productModel));
                             }
                         });
+                if (isNotIdExist.get()) {
+                    appendError("User was not found!");
+                    return false;
+                }
                 Files.write(Paths.get(FILE_NAME), products);
                 return true;
             } catch (Throwable t) {
